move batchTotals helper from geBatchGiftEntryHeader to geGiftEntryFormApp
