+++ src/lwc/geGiftEntryFormApp/geGiftEntryFormApp.html
@@ -2,11 +2,11 @@
     <template if:true={isBatchMode}>
         <c-ge-batch-gift-entry-header
                 batch-id={recordId}
+                batch-totals={batchTotals}
                 is-permission-error={isPermissionError}
                 onbatchdryrun={handleBatchDryRun}
                 onedit={handleEditBatch}
-                onprocessbatch={handleProcessBatch}
-                ontogglemodal={handleToggleModal}>
+                onprocessbatch={handleProcessBatch}>
         </c-ge-batch-gift-entry-header>
         <br/>
     </template>
@@ -37,7 +37,7 @@
                     oncountchanged={handleCountChanged}
                     ontotalchanged={handleTotalChanged}
                     onselectcolumns={handleSelectColumns}
-                    batch-currency-iso-code= {batchCurrencyIsoCode}>
+                    batch-currency-iso-code={batchCurrencyIsoCode}>
             </c-ge-batch-gift-entry-table>
         </template>
     </template>
