+++ src/lwc/geGiftEntryFormApp/helpers/batchTotals.js
